Added primitives and operations
