+++ sdf_geometry/operations.py
@@ -1,0 +1,5 @@
+import numpy as np
+
+def smooth_union(a, b, k=10.0):
+    return -np.log(np.exp(-k * a) + np.exp(-k * b)) / k
+
